Move to package structure
* Copy paste
* Fixed imports
* Test file only moved!
