+++ setup.cfg
@@ -48,7 +48,8 @@ universal = 0
 test = pytest
 
 [tool:pytest]
-addopts = --pycodestyle -v --cov --cov-report term-missing
+testpaths = tests
+addopts = --pycodestyle -v --cov piplicenses --cov-report term-missing
 
 [tool:isort]
 # https://github.com/timothycrosley/isort
