+++ tests/test_piplicenses.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-# vim:fenc=utf-8 ff=unix ft=python ts=4 sw=4 sts=4 si et
 import copy
 import re
 import sys
